Refactoring the SMT folder
